+++ src/Classes/Student.java
@@ -1,14 +1,12 @@
 package Classes;
 
-import java.math.BigInteger;
-
 public class Student extends Person{
     private boolean present; // today's attendance state
     private int past; // number of times student was absent
     private String Excuse; // excuse for last time student was absent
 
     // a constructor with default excuse message
-    public Student(String name, BigInteger ID, boolean present, int past) {
+    public Student(String name, int ID, boolean present, int past) {
         this.name = name;
         this.ID = ID;
         this.present = present;
@@ -17,7 +15,7 @@ public class Student extends Person{
     }
 
     // a constructor with last excuse
-    public Student(String name, BigInteger ID, boolean present, int past, String excuse) {
+    public Student(String name, int ID, boolean present, int past, String excuse) {
         this.name = name;
         this.ID = ID;
         this.present = present;
